camera_follows_player_in_frozen_time
As a player, when I die, I want to visually see my character die and come back to life to feel that I have died. (120mins)
As a player, when I respawn, I want to visually see my character come back to life to feel that I have respawned. (120 mins)